requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
assertpy==1.1
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
iniconfig==1.1.1
jsonschema==3.2.0
packaging==20.9
paramiko==2.7.2
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
pytest-html==3.1.1
pytest-metadata==1.11.0
requests==2.25.1
six==1.15.0
sshtunnel==0.4.0
toml==0.10.2
urllib3==1.26.4
cerberus==1.3.3
PyMySQL==1.0.2
invoke==1.5.0
pytest-rerunfailures==9.1.1